+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/impl/CaseOrchestrationServiceImpl.java
@@ -925,9 +925,4 @@ public class CaseOrchestrationServiceImpl implements CaseOrchestrationService {
                 .build();
         }
     }
-
-    @Override
-    public Map<String, Object> receivedServiceAddedDate(CcdCallbackRequest ccdCallbackRequest) throws WorkflowException {
-        return receivedServiceAddedDateWorkflow.run(ccdCallbackRequest.getCaseDetails());
-    }
 }
