+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/impl/ServiceJourneyServiceImpl.java
@@ -4,10 +4,14 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CcdCallbackRequest;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CcdCallbackResponse;
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.WorkflowException;
 import uk.gov.hmcts.reform.divorce.orchestration.service.ServiceJourneyService;
 import uk.gov.hmcts.reform.divorce.orchestration.workflows.MakeServiceDecisionDateWorkflow;
+import uk.gov.hmcts.reform.divorce.orchestration.workflows.ReceivedServiceAddedDateWorkflow;
+
+import java.util.Map;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdStates.AWAITING_DN_APPLICATION;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdStates.SERVICE_APPLICATION_NOT_APPROVED;
@@ -20,6 +24,7 @@ import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.Orchestrati
 public class ServiceJourneyServiceImpl implements ServiceJourneyService {
 
     private final MakeServiceDecisionDateWorkflow makeServiceDecisionDateWorkflow;
+    private final ReceivedServiceAddedDateWorkflow receivedServiceAddedDateWorkflow;
 
     @Override
     public CcdCallbackResponse makeServiceDecision(CaseDetails caseDetails) throws WorkflowException {
@@ -34,6 +39,12 @@ public class ServiceJourneyServiceImpl implements ServiceJourneyService {
         builder.data(makeServiceDecisionDateWorkflow.run(caseDetails));
 
         return builder.build();
+    }
+
+    @Override
+    public Map<String, Object> receivedServiceAddedDate(CcdCallbackRequest ccdCallbackRequest)
+        throws WorkflowException {
+        return receivedServiceAddedDateWorkflow.run(ccdCallbackRequest.getCaseDetails());
     }
 
     protected boolean isServiceApplicationGranted(CaseDetails caseDetails) {
